Restructure: move more to shared
