+++ packages/scandipwa/packages/runtime-i18n/build-config/webpack-i18n-import-loader/index.js
@@ -4,7 +4,7 @@ const loaderUtils = require('loader-utils');
 
 const parentThemeHelper = require('@scandipwa/scandipwa-dev-utils/parent-theme');
 const extensions = require('@scandipwa/scandipwa-dev-utils/extensions');
-const getEnabledLocales = require('../shared/getEnabledLocales');
+const getEnabledLocales = require('../../shared/getEnabledLocales');
 
 const childRoot = process.cwd();
 const parentRoots = parentThemeHelper.getParentThemePaths(childRoot);
